multithreading working + updated launch script

+++ scripts/launch_script.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+cd ../build
+
+if ! make; then
+    exit 1
+fi
+
 tmux kill-window -t +1
 tmux new-window -t $SESSION:+1 -c ./bin
 
@@ -7,11 +13,11 @@ tmux split-window -h -c ./bin
 tmux split-window -h -c ./bin
 
 tmux select-pane -t 0
-tmux send-keys "cd ../build/ && make && cd ../bin && ./binder > outfile" Enter
+tmux send-keys "./binder > outfile" Enter
 
 tmux select-pane -t 1
-tmux send-keys "sleep 2 && source ../scripts/get_vars.sh && ./client" Enter
+tmux send-keys "sleep 1 && source ../scripts/get_vars.sh && ./client" Enter
 
 tmux select-pane -t 2
-tmux send-keys "sleep 1 && source ../scripts/get_vars.sh && ./server" Enter
+tmux send-keys "sleep 0.5 && source ../scripts/get_vars.sh && ./server" Enter
